Added 'Era' Suffix; Remove '.md' from title
## Era Display
Added the "era" data field to the span block. If specified, this will be appended to the date in the timeline display.

For example, this span would display in the timeline as "144-43-49-00 AB".
```
<span
	  class='ob-timelines'
	  data-date='144-43-49-00'
     data-era="AB"
>
```

## Remove `.md` from autofilled titles
When auto-filling the title based on the filename, remove the `.md` from the end. More in line with how other linking/embedding is displayed through obsidian.

Also updated the `package.json` to the latest standard for including the obsidian API. See https://github.com/obsidianmd/obsidian-api/issues/17.